Move experimental attention and residual features
