+++ src/__init__.py
@@ -7,9 +7,6 @@ from .modeling import (
     BiBoAttention,
     BiBoMLP,
     BiBoMoELayer,
-    BiBoResidualGate,
-    BiBoCausalResidualConv,
-    BiBoMultiStreamResidual,
     BiBoRMSNorm,
 )
 
@@ -21,8 +18,5 @@ __all__ = [
     'BiBoAttention',
     'BiBoMLP',
     'BiBoMoELayer',
-    'BiBoResidualGate',
-    'BiBoCausalResidualConv',
-    'BiBoMultiStreamResidual',
     'BiBoRMSNorm',
 ]
