+++ src/exp/__init__.py
@@ -1,0 +1,5 @@
+"""Experimental BiBo components.
+
+Stable model code imports from this package only when an experimental config
+option is enabled. Canonical experimental config lives under ``BiBoConfig(exp=...)``.
+"""
