+++ src/exp/attn/__init__.py
@@ -1,0 +1,5 @@
+"""Experimental attention variants."""
+
+from .recurrent import *
+from .sliding import *
+from .ssmax import *
