+++ src/modeling/attn/__init__.py
@@ -1,7 +1,4 @@
 """Attention components"""
 from .base import *
 from .standard import *
-from .sliding import *
-from .recurrent import *
-from .ssmax import *
 from .utils import *
